Real ExpressionTask and EquationTask

+++ lab-01/by/aadeglmmy/quizer/tasks/math_tasks/MathTask.java
@@ -10,11 +10,11 @@ public interface MathTask extends Task {
 
   interface Generator extends Task.Generator {
 
-    int getMinNumber();
+    double getMinNumber();
 
-    int getMaxNumber();
+    double getMaxNumber();
 
-    default int getDiffNumber() {
+    default double getDiffNumber() {
       return getMaxNumber() - getMinNumber();
     }
   }
